Optimizations

See merge request ls11ae/research/random-walks!6

+++ src/cuda/brownian_gpu.h
@@ -11,9 +11,10 @@
 extern "C" {
 #endif
 
-Point2DArray *gpu_brownian_walk(Matrix *kernel_matrix, size_t T, size_t W, size_t H, size_t start_x, size_t start_y,
-                                size_t end_x,
-                                size_t end_y);
+Point2DArray *gpu_brownian_walk(Matrix *kernel_matrix, uint32_t T, uint32_t W, uint32_t H, uint32_t start_x,
+                                uint32_t start_y,
+                                uint32_t end_x,
+                                uint32_t end_y);
 
 #ifdef __cplusplus
 }
