+++ src/cuda/cuda_adapter.h
@@ -28,13 +28,13 @@ typedef struct {
 
 Tensor *tensor_new_empty(int D);
 
-void tensor_flat(const Tensor *t, double *values);
+void tensor_flat(const Tensor *t, float *values);
 
-Tensor *tensor_from_flat(const double *flat, size_t tensor_len, ssize_t mat_width, ssize_t mat_height);
+Tensor *tensor_from_flat(const float *flat, uint32_t tensor_len, int32_t mat_width, int32_t mat_height);
 
-void dir_kernel_to_cuda(const Vector2D *input, int2 **out_offsets, int **out_sizes, size_t *out_D);
+void dir_kernel_to_cuda(const Vector2D *input, int2 **out_offsets, int **out_sizes, uint32_t *out_D);
 
-Tensor **convert_dp_host_to_tensor(const double *dp_host, ssize_t T, ssize_t D, ssize_t H, ssize_t W);
+Tensor **convert_dp_host_to_tensor(const float *dp_host, int32_t T, int32_t D, int32_t H, int32_t W);
 
 #ifdef __cplusplus
 }
