+++ src/parsers/types.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <stdbool.h>
-#include <stddef.h>
 #include <sys/types.h>
 #include <stdint.h>
 #include <linux/limits.h>
@@ -15,121 +14,121 @@ extern "C" {
  * @brief Represents a 2D matrix.
  */
 typedef struct {
-    ssize_t width; /**< The number of columns in the matrix. */
-    ssize_t height; /**< The number of rows in the matrix. */
-    ssize_t len; /**< The total number of elements (width * height). */
-    double* data; /**< Pointer to the array of matrix elements. */
+    int32_t width; /**< The number of columns in the matrix. */
+    int32_t height; /**< The number of rows in the matrix. */
+    int32_t len; /**< The total number of elements (width * height). */
+    float *data; /**< Pointer to the array of matrix elements. */
 } Matrix;
 
 typedef struct {
-    ssize_t x;
-    ssize_t y;
+    int32_t x;
+    int32_t y;
 } Point2D;
 
 typedef struct {
-    Point2D** data;
-    size_t* sizes;
-    size_t count;
+    Point2D **data;
+    uint32_t *sizes;
+    uint32_t count;
 } Vector2D;
 
 typedef struct {
-    //size_t dim_len;
-    //size_t *dim;
-    size_t len;
-    Matrix** data;
-    Vector2D* dir_kernel;
+    //uint32_t dim_len;
+    //uint32_t *dim;
+    uint32_t len;
+    Matrix **data;
+    Vector2D *dir_kernel;
 } Tensor;
 
 
 #define HASH_CACHE_BUCKETS 4096
 
 typedef struct HashEntry {
-    uint64_t hash;
-    Tensor* tensor;
+    uint32_t hash;
+    Tensor *tensor;
     char path[PATH_MAX];
-    struct HashEntry* next;
+    struct HashEntry *next;
 } HashEntry;
 
 typedef struct HashCache {
-    HashEntry* buckets[HASH_CACHE_BUCKETS];
+    HashEntry *buckets[HASH_CACHE_BUCKETS];
 } HashCache;
 
 typedef struct CacheEntry {
-    uint64_t hash;
+    uint32_t hash;
 
     union {
-        Tensor* array; // For tensor_map_new
-        Matrix* single; // For kernels_map_new
+        Tensor *array; // For tensor_map_new
+        Matrix *single; // For kernels_map_new
     } data;
 
     bool is_array;
-    ssize_t array_size;
-    struct CacheEntry* next;
+    int32_t array_size;
+    struct CacheEntry *next;
 } CacheEntry;
 
 typedef struct {
-    CacheEntry** buckets;
-    size_t num_buckets;
+    CacheEntry **buckets;
+    uint32_t num_buckets;
 } Cache;
 
 typedef struct {
-    //size_t dim_len;
-    //size_t *dim;
-    size_t len;
-    size_t max_D;
-    Tensor** data;
-    Vector2D** grid_cells;
+    //uint32_t dim_len;
+    //uint32_t *dim;
+    uint32_t len;
+    uint32_t max_D;
+    Tensor **data;
+    Vector2D **grid_cells;
 } TensorSet;
 
 typedef struct {
-    Point2D* points;
-    size_t length;
+    Point2D *points;
+    uint32_t length;
 } Point2DArray;
 
 typedef struct {
-    Point2DArray*** data;
-    size_t width;
-    size_t height;
-    size_t times;
+    Point2DArray ***data;
+    uint32_t width;
+    uint32_t height;
+    uint32_t times;
 } Point2DArrayGrid;
 
 typedef struct {
-    double x; // longitude
-    double y; // latitude
+    float x; // longitude
+    float y; // latitude
 } Coordinate;
 
 typedef struct {
-    Coordinate* points;
-    size_t length;
+    Coordinate *points;
+    uint32_t length;
 } Coordinate_array;
 
 
 typedef struct {
     bool is_brownian;
-    ssize_t S;
-    ssize_t D;
+    int32_t S;
+    int32_t D;
     float diffusity;
-    ssize_t bias_x;
-    ssize_t bias_y;
+    int32_t bias_x;
+    int32_t bias_y;
 } KernelParameters;
 
 
 typedef struct {
-    Matrix*** kernels;
-    ssize_t width, height;
-    Cache* cache;
+    Matrix ***kernels;
+    int32_t width, height;
+    Cache *cache;
 } KernelsMap;
 
 typedef struct {
-    Tensor*** kernels; // 3D [y][x][d]
-    ssize_t width, height, max_D;
-    Cache* cache;
+    Tensor ***kernels; // 3D [y][x][d]
+    int32_t width, height, max_D;
+    Cache *cache;
 } KernelsMap3D;
 
 typedef struct {
-    Tensor**** kernels; // 4D array [y][x][t][d]
-    ssize_t width, height, timesteps, max_D;
-    Cache* cache;
+    Tensor ****kernels; // 4D array [y][x][t][d]
+    int32_t width, height, timesteps, max_D;
+    Cache *cache;
 } KernelsMap4D;
 
 enum landmarkType {
@@ -148,16 +147,16 @@ enum landmarkType {
 
 
 typedef struct {
-    size_t width;
-    size_t height;
-    KernelParameters*** data;
+    uint32_t width;
+    uint32_t height;
+    KernelParameters ***data;
 } KernelParametersTerrain;
 
 typedef struct {
-    size_t width;
-    size_t height;
-    size_t time;
-    KernelParameters**** data;
+    uint32_t width;
+    uint32_t height;
+    uint32_t time;
+    KernelParameters ****data;
 } KernelParametersTerrainWeather;
 
 typedef struct {
@@ -172,19 +171,19 @@ typedef struct {
 } WeatherEntry;
 
 typedef struct {
-    WeatherEntry** data;
-    size_t length;
+    WeatherEntry **data;
+    uint32_t length;
 } WeatherTimeline;
 
 typedef struct {
-    size_t height;
-    size_t width;
-    WeatherTimeline*** entries; // Timeline at [y][x]
+    uint32_t height;
+    uint32_t width;
+    WeatherTimeline ***entries; // Timeline at [y][x]
 } WeatherGrid;
 
 typedef struct {
-    int** data;
-    ssize_t width, height;
+    int **data;
+    int32_t width, height;
 } TerrainMap;
